+++ lottery/types.go
@@ -1,4 +1,4 @@
-package main
+package lottery
 
 import "time"
