+++ main.go
@@ -19,33 +19,10 @@ import (
 	"strconv"
 
 	_ "github.com/mattn/go-sqlite3"
+	"./handler"
 )
 
-type Handler struct{
-}
 
-func (h Handler)RootHandler(w http.ResponseWriter, r *http.Request){
-	resp, err := http.Get("https://lottery-dot-tenntenn-samples.appspot.com/available_lotteries")
-	if err != nil {
-		const status = http.StatusInternalServerError
-		http.Error(w, http.StatusText(status), status)
-		return
-	}
-	defer resp.Body.Close()
-
-	var lotteries []*Lottery
-	if err := json.NewDecoder(resp.Body).Decode(&lotteries); err != nil {
-		const status = http.StatusInternalServerError
-		http.Error(w, http.StatusText(status), status)
-		return
-	}
-
-	if err := listTmpl.Execute(w, lotteries); err != nil {
-		const status = http.StatusInternalServerError
-		http.Error(w, http.StatusText(status), status)
-		return
-	}
-}
 
 func main() {
 
@@ -58,7 +35,7 @@ func main() {
 		panic(err)
 	}
 
-	var h Handler
+	var h handler.Handler
 
 	http.HandleFunc("/", h.RootHandler)
